requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

onnx>=1.16.0
onnxruntime==1.17.1

numpy==1.26.4
pillow==10.2.0

python-multipart==0.0.9
joblib==1.4.2
